Access `TyCtxt` from early diagnostic decoration
